monitoring/discord-cs-bot-liam/requirements.txt
discord.py
python-dotenv
requests
beautifulsoup4
cloudscraper
selenium
webdriver-manager